Introduce AbstractFilePropertyMetadatFactory to remove duplicated code.
